Counting test files for merge
Counting files for merge.
Adding some TODOs for merge testing.
Moving the countfiles code to a common file.
