+++ lib/merge_test.go
@@ -7,6 +7,12 @@ import (
 	"testing"
 )
 
+/*
+Think of all test cases:
+* Merge with bad named path to make invalid tree.
+* Merge with collision files
+* Merge with different files
+*/
 func TestMergeCheckGood(t *testing.T) {
 	for method := MethodYear; method < MethodNone; method++ {
 		td := newTestDir(t, method)
@@ -18,8 +24,7 @@ func TestMergeCheckGood(t *testing.T) {
 
 		err := MergeCheck(dst, method)
 		if err != nil {
-			t.Errorf("Err %s, method %d\n",
-				err.Error(), method)
+			t.Errorf("Err %s, method %d\n", err.Error(), method)
 		}
 
 		os.RemoveAll(dst)
@@ -55,17 +60,32 @@ func TestMergeDuplicate(t *testing.T) {
 		td := newTestDir(t, method)
 
 		src := td.buildRoot()
-		defer os.RemoveAll(src)
 
-		dst1 := td.buildSortedDir(src, "dst1_", ActionCopy)
-		dst2 := td.buildSortedDir(src, "dst2_", ActionCopy)
+		// COpy files to two sorted directories that are identical
+		fromDir := td.buildSortedDir(src, "fromDir_", ActionCopy)
+		toDir := td.buildSortedDir(src, "toDir_", ActionCopy)
 
-		err := Merge(dst1, dst2, ActionMove, ioutil.Discard)
+		// merge them
+		err := Merge(fromDir, toDir, ActionMove, ioutil.Discard)
 		if err != nil {
 			t.Errorf("Error %s", err.Error())
 		}
 
-		os.RemoveAll(dst1)
-		os.RemoveAll(dst2)
+		// Source should have all its files removed as they are all
+		// duplicates
+		err = countFiles(t, fromDir, 0, "Src Dir")
+		if err != nil {
+			t.Errorf("Error %s", err.Error())
+		}
+
+		// Destination should have all its original contents
+		err = countFiles(t, toDir, td.numTotal(), "Target Dir")
+		if err != nil {
+			t.Errorf("Error %s", err.Error())
+		}
+
+		os.RemoveAll(fromDir)
+		os.RemoveAll(toDir)
+		os.RemoveAll(src)
 	}
 }
