+++ lib/testdir_test.go
@@ -1,6 +1,7 @@
 package exifsort
 
 import (
+	"errors"
 	"fmt"
 	"io/ioutil"
 	"os"
@@ -54,6 +55,36 @@ const (
 	skipPath       = "../README.md"
 	nonesensePath  = "../gobofragggle"
 )
+
+func countFiles(t *testing.T, path string, correctCount int, label string) error {
+	var count = 0
+
+	_ = filepath.Walk(path,
+		func(path string, info os.FileInfo, err error) error {
+			if err != nil {
+				count++
+				return nil
+			}
+
+			if info.IsDir() {
+				return nil
+			}
+
+			count++
+
+			return nil
+		})
+
+	if count != correctCount {
+		errStr := fmt.Sprintf("count error for %s on %s. "+
+			"Expected %d got %d",
+			label, path, correctCount, count)
+
+		return errors.New(errStr)
+	}
+
+	return nil
+}
 
 type testdir struct {
 	fileNo    int
